Adds option to cycle out workers
This adds an option to cycle out workers after a given expiration. The
idea is that a worker (a phantom process) will not stay alive for too
long of a time. If you set `expiration` in your options to a millisecond
value we'll cycle out workers after that amount of time. Fixes #13